reuse the existing box_ops
